Made changes to disable path mtu discovery and IP_HDRINCL for raw socket. Also added code to tweak poll_watermark and poll_duration of pfring. This needed adding code to gopacket library itself. This tweaking let to a good increase in performance.

+++ cmd/dataplane/fib/lisp.go
@@ -51,7 +51,7 @@ func SetLispIID(hdr []byte, iid uint32) {
 func GetLispIID(hdr []byte) uint32 {
 	var iid uint32 = 0
 
-	iid = uint32(hdr[4]<<16 | hdr[5]<<8 | hdr[6])
+	iid = uint32(uint32(hdr[4])<<16 | uint32(hdr[5])<<8 | uint32(hdr[6]))
 	return iid
 }
